refactor(ui): split App.tsx into shell + page registry + nav groups
App.tsx ballooned to 639 lines and was tripping the project's red-flag
size budget. Most of that was per-page chrome metadata (route table
with title / description / icon / lazy-loaded component / contextual
help body) that has nothing to do with the top-level shell. Splitting
by concern makes each file boring in isolation and means adding a new
page is a single-file edit.

Layout:
- src/App.tsx (131)         — shell: ErrorBoundary -> AppProvider ->
                              AppShell, plus the route table that
                              consumes pages[] and the dynamic
                              /device-config/* routes
- src/pageRegistry.tsx (487) — pages[] route table, the PageConfig
                              type, and the lazy() imports for every
                              non-eager page (kept colocated so the
                              entry, the lazy import, and the metadata
                              never drift apart)
- src/navGroups.ts (75)     — sidebar navGroups (Control / Configuration
                              / Network / Analysis / Automation)
- src/ui/PageLoader.tsx (15) — Suspense fallback for lazy chunks

Pure code move. Behaviour is unchanged: same routes, same component
references, same lazy-load boundaries, same error boundary keying on
useLocation().pathname. biome / tsc / vitest all pass; make build
produces a working bundle.
